human activity folder restructuring
